AES.key: update LM14 key comment
The LM14 key is used for OTAIDs HE_LCD_NC5U_AAADABAA and
HE_LCD_NC5B_AFADABAA, which are both NetCast 4.5. Based on the filename
of an HE_LCD_NC5B_AFADABAA firmware image, it seems like it may be for
2015 models.

Also indicate that not all LH56* models use SimpleSmart.